Added config option for centering the compass text on screen
All variables, fields, and methods (with the exceptions of class/interface/enum names, constructors, and some static constants) should now be in camelCase, with the first letter being lowercase. All static constant exceptions to this should be in all uppercase, including enum constants. Constructors match their class name, in which the first letter is uppercase, but otherwise matches the camelCase used everywhere else.

Took me the better part of the day, but I feel its worth it.